refactor: 정보아이콘으로 변경 및 코드 스타일 수정
