+++ src/components/Popover/index.tsx
@@ -1,23 +1,27 @@
-'use client';
+import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react';
 
-import * as React from 'react';
-
-import * as PopoverPrimitive from '@radix-ui/react-popover';
+import {
+  Anchor,
+  Content,
+  Portal,
+  Root,
+  Trigger,
+} from '@radix-ui/react-popover';
 
 import { cn } from '@/utils/cn';
 
-const Popover = PopoverPrimitive.Root;
+const Popover = Root;
 
-const PopoverTrigger = PopoverPrimitive.Trigger;
+const PopoverTrigger = Trigger;
 
-const PopoverAnchor = PopoverPrimitive.Anchor;
+const PopoverAnchor = Anchor;
 
-const PopoverContent = React.forwardRef<
-  React.ElementRef<typeof PopoverPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Content>
+const PopoverContent = forwardRef<
+  ElementRef<typeof Content>,
+  ComponentPropsWithoutRef<typeof Content>
 >(({ className, align = 'center', sideOffset = 4, ...props }, ref) => (
-  <PopoverPrimitive.Portal>
-    <PopoverPrimitive.Content
+  <Portal>
+    <Content
       ref={ref}
       align={align}
       sideOffset={sideOffset}
@@ -27,8 +31,8 @@ const PopoverContent = React.forwardRef<
       )}
       {...props}
     />
-  </PopoverPrimitive.Portal>
+  </Portal>
 ));
-PopoverContent.displayName = PopoverPrimitive.Content.displayName;
+PopoverContent.displayName = Content.displayName;
 
 export { Popover, PopoverTrigger, PopoverContent, PopoverAnchor };
